Adds projectID to integration test flags
