+++ edge-node-simulator/test/README.md
@@ -18,6 +18,7 @@ ONBPASS="" # The orch keycloak user password - to retrieve token for Infrastruct
 APIUSER="" # The orch keycloak user - to retrieve token for Infrastructure Manager REST API interactions - if not specified goes to default
 APIPASS="" # The orch keycloak user password - to retrieve token for Infrastructure Manager REST API interactions - if not specified goes to default
 PROJECT="" # The project name in which the ONBUSER and APIUSER belong to.
+PROJECTID="" # The project ID in which the APIUSER and APIPASS belong to.
 ```
 
 ## Edge Node Simulator Deployment
@@ -48,7 +49,7 @@ kubectl port-forward svc/api -n orch-infra --address 0.0.0.0 8080:8080 &
 
 ```bash
 ginkgo -v -r --fail-fast --race --json-report infra-tests-day0.json --output-dir . --label-filter="infra-tests-day0" ./test/infra -- \
-    -project=${PROJECT} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
+    -project=${PROJECT} -projectID=${PROJECTID} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
     -orchFQDN=${ORCH_FQDN} -infraURL=${API_URL} \
     -edgeAPIUser=${APIUSER}  -edgeAPIPass=${APIPASS} \
     -edgeOnboardUser=${ONBUSER} -edgeOnboardPass=${ONBPASS}
@@ -58,7 +59,7 @@ ginkgo -v -r --fail-fast --race --json-report infra-tests-day0.json --output-dir
 
 ```bash
 ginkgo -v -r --fail-fast --race --json-report infra-tests-day1.json --output-dir . --label-filter="infra-tests-day1" ./test/infra -- \
-    -project=${PROJECT} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
+    -project=${PROJECT} -projectID=${PROJECTID} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
     -orchFQDN=${ORCH_FQDN} -infraURL=${API_URL} \
     -edgeAPIUser=${APIUSER}  -edgeAPIPass=${APIPASS} \
     -edgeOnboardUser=${ONBUSER} -edgeOnboardPass=${ONBPASS}
@@ -68,7 +69,7 @@ ginkgo -v -r --fail-fast --race --json-report infra-tests-day1.json --output-dir
 
 ```bash
 ginkgo -v -r --fail-fast --race --json-report infra-tests-day2.json --output-dir . --label-filter="infra-tests-day2" ./test/infra --  \
-    -project=${PROJECT} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
+    -project=${PROJECT} -projectID=${PROJECTID} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
     -orchFQDN=${ORCH_FQDN} -infraURL=${API_URL} \
     -edgeAPIUser=${APIUSER}  -edgeAPIPass=${APIPASS} \
     -edgeOnboardUser=${ONBUSER} -edgeOnboardPass=${ONBPASS}
@@ -78,7 +79,7 @@ ginkgo -v -r --fail-fast --race --json-report infra-tests-day2.json --output-dir
 
 ```bash
 ginkgo -v -r --fail-fast --race --label-filter="cleanup" ./test/infra --  \
-    -project=${PROJECT} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
+    -project=${PROJECT} -projectID=${PROJECTID} -caFilepath=${CA_PATH} -simAddress=${ENSIM_ADDR} \
     -orchFQDN=${ORCH_FQDN} -infraURL=${API_URL} \
     -edgeAPIUser=${APIUSER}  -edgeAPIPass=${APIPASS} \
     -edgeOnboardUser=${ONBUSER} -edgeOnboardPass=${ONBPASS}
